requirements.txt
astroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==2.2.2
python-engineio==3.2.1
python-socketio==3.1.1
requests==2.21.0
requirements==0.1
six==1.12.0
typed-ast==1.1.1
urllib3==1.24.1
wrapt==1.10.11
